Move interfaces to common
