+++ src/app/client/ScrcpyClient.ts
@@ -1,25 +1,16 @@
 import MseDecoder from '../decoder/MseDecoder';
 import { DeviceController } from '../DeviceController';
 import BroadwayDecoder from '../decoder/BroadwayDecoder';
-import { ParsedUrlQueryInput } from 'querystring';
 import { BaseClient } from './BaseClient';
 import Decoder from '../decoder/Decoder';
 import Tinyh264Decoder from '../decoder/Tinyh264Decoder';
-
-export type Decoders = 'broadway' | 'mse' | 'tinyh264';
-
-export interface StreamParams extends ParsedUrlQueryInput {
-    action: 'stream';
-    udid: string;
-    decoder: Decoders;
-    ip: string;
-    port: string;
-}
+import { Decoders } from '../../common/Decoders';
+import { ScrcpyStreamParams } from '../../common/ScrcpyStreamParams';
 
 export class ScrcpyClient extends BaseClient {
     public static ACTION = 'stream';
     private static instance?: ScrcpyClient;
-    public static start(params: StreamParams): ScrcpyClient {
+    public static start(params: ScrcpyStreamParams): ScrcpyClient {
         const client = this.getInstance();
         client.startStream(params.udid, params.decoder, `ws://${params.ip}:${params.port}`);
         client.setBodyClass('stream');
