+++ src/common/Decoders.d.ts
@@ -1,0 +1,1 @@
+export type Decoders = 'broadway' | 'mse' | 'tinyh264';
