+++ src/common/ShellParams.d.ts
@@ -1,0 +1,6 @@
+import { ParsedUrlQueryInput } from 'querystring';
+
+export interface ShellParams extends ParsedUrlQueryInput {
+    action: 'shell';
+    udid: string;
+}
